# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Keranjang Service API",
        "description": "Dokumentasi API untuk Keranjang Service pada sistem E-Commerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/carts": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Mengambil seluruh data keranjang",
                "operationId": "4596e9e6e6d3b2273ffbaa1254d0303f",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Invalid or missing API Key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/carts/{id}": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Mengambil detail keranjang",
                "operationId": "6f42ea598ee06538f7528d2e39acfd7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Cart not found"
                    },
                    "401": {
                        "description": "Invalid or missing API Key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/carts/items": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "Menambahkan produk ke keranjang",
                "operationId": "b5c832dd22b0a836b277080d56ddbb8c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cart_id",
                                    "product_id",
                                    "product_name",
                                    "quantity",
                                    "price"
                                ],
                                "properties": {
                                    "cart_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "product_name": {
                                        "type": "string",
                                        "example": "Webcam HD"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 180000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Invalid or missing API Key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Carts",
            "description": "Carts"
        }
    ]
}